requirements.txt
flask
boto3
pymongo
python-dotenv
pyjwt
twilio
gunicorn
flask-cors
dnspython